feat (Shopping cart): add shopping cart view
